Move result_parser to parties app, and remove demoscene.utils
