+++ parties/result_parser.py
